refactor: move sso-mib token fetching into separate compilation unit
